Upgrade from version 0.62a to cvs version now work correctly on postgresql FIX bug #4821 "Pb de suppression des photos" ADD fix from bug #4821 to preferences.php

+++ ajouter_adherent.php
@@ -273,7 +273,7 @@
               $error_detected[] = _T("Update Blob failed");
 					}
 				}
-        if (isset($_POST['del_photo']))
+        if ( isset($_POST['del_photo']) && $_POST['del_photo'] == 1 )
         {
           $sql = "DELETE FROM ".PREFIX_DB."pictures
             WHERE id_adh=".$adherent['id_adh'];
